Deletes model/init.js and moves table creation query to database/init.js, successfully creates tables
