+++ server/index.js
@@ -9,5 +9,5 @@ const port = 3000;
 app.listen(port, () => {
   console.log('[Server] Now listening on port', port);
 
-  require('./model/photos.js');
+  require('./database/init.js');
 });
